Move conversion function back to logger
